requirements.txt
# Agent frameworks
crewai[tools]>=0.86.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0

# Vector DB & RAG
qdrant-client>=1.12.0
sentence-transformers>=3.3.0
langchain-text-splitters>=0.3.0

# Russian NLP
pymorphy3>=2.0
pymorphy3-dicts-ru>=2.0

# Web search
duckduckgo-search>=7.0.0

# API
fastapi>=0.115.0
uvicorn>=0.34.0
python-multipart>=0.0.18

# UI
streamlit>=1.41.0

# Config
pydantic-settings>=2.7.0
PyYAML>=6.0

# Utilities
httpx>=0.28.0

# Database (content factory)
psycopg2-binary>=2.9.0

# Source collectors
praw>=7.8.0
feedparser>=6.0.0

# Scheduler
apscheduler>=3.10.0

# Charts
matplotlib>=3.9.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
